+++ package.json
@@ -17,7 +17,7 @@
     "build:db-postgres": "turbo build --filter db-postgres",
     "build:email-nodemailer": "turbo build --filter email-nodemailer",
     "build:email-resend": "turbo build --filter email-resend",
-    "build:eslint-config-payload": "turbo build --filter eslint-config-payload",
+    "build:eslint-config": "turbo build --filter eslint-config",
     "build:force": "pnpm run build:core:force",
     "build:graphql": "turbo build --filter graphql",
     "build:live-preview": "turbo build --filter live-preview",
@@ -93,6 +93,7 @@
     "@jest/globals": "29.7.0",
     "@next/bundle-analyzer": "^14.1.0",
     "@payloadcms/eslint-config": "workspace:*",
+    "@payloadcms/eslint-plugin": "workspace:*",
     "@payloadcms/live-preview-react": "workspace:*",
     "@playwright/test": "1.43.0",
     "@swc/cli": "0.3.12",
@@ -117,7 +118,6 @@
     "drizzle-kit": "0.20.14-1f2c838",
     "drizzle-orm": "0.29.4",
     "escape-html": "^1.0.3",
-    "eslint-plugin-payload": "workspace:*",
     "execa": "5.1.1",
     "form-data": "3.0.1",
     "fs-extra": "10.1.0",
