+++ packages/eslint-config/configs/jest/index.mjs
