+++ packages/eslint-config/configs/react/index.mjs
