+++ packages/eslint-config/configs/react/rules/react.mjs
