+++ packages/eslint-plugin/index.mjs
