+++ test/package.json
@@ -20,6 +20,7 @@
     "@payloadcms/email-nodemailer": "workspace:*",
     "@payloadcms/email-resend": "workspace:*",
     "@payloadcms/eslint-config": "workspace:*",
+    "@payloadcms/eslint-plugin": "workspace:*",
     "@payloadcms/graphql": "workspace:*",
     "@payloadcms/live-preview": "workspace:*",
     "@payloadcms/live-preview-react": "workspace:*",
@@ -50,7 +51,6 @@
     "comment-json": "^4.2.3",
     "create-payload-app": "workspace:*",
     "dotenv": "16.4.5",
-    "eslint-plugin-payload": "workspace:*",
     "eslint-plugin-playwright": "1.6.2",
     "execa": "5.1.1",
     "http-status": "1.6.2",
